Fix trig functions

+++ Primitives/camera.cpp
@@ -13,15 +13,15 @@ namespace Primitives {
 
     algebra::Matrix4f Camera::GetViewMatrix() {
         algebra::Vector3f rot = algebra::Vector3f({
-            std::atan2f(lookat.y(), lookat.z()),
-            -std::atan2f(lookat.x(), lookat.z()),
+            std::atan2(lookat.y(), lookat.z()),
+            -std::atan2(lookat.x(), lookat.z()),
             0.0
         });
 
         algebra::Matrix4f rot_matrix = algebra::Matrix4f({
-            std::cosf(rot.y()) * std::cosf(rot.z()), std::sinf(rot.x()) * std::sinf(rot.y()) * std::cosf(rot.z()) - std::cosf(rot.x()) * std::sinf(rot.z()), std::cosf(rot.x()) * std::sinf(rot.y()) * std::cosf(rot.z()) + std::sinf(rot.x()) * std::sinf(rot.z()), 0.0f,
-            std::cosf(rot.y()) * std::sinf(rot.z()), std::sinf(rot.x()) * std::sinf(rot.y()) * std::sinf(rot.z()) + std::cosf(rot.x()) * std::cosf(rot.z()), std::cosf(rot.x()) * std::sinf(rot.y()) * std::sinf(rot.z()) - std::sinf(rot.x()) * std::cosf(rot.z()), 0.0f,
-            -std::sinf(rot.y()),                     std::sinf(rot.x()) * std::cos(rot.y()),                                                                 std::cosf(rot.x()) * std::cosf(rot.y()),                                                                0.0f,
+            std::cos(rot.y()) * std::cos(rot.z()), std::sin(rot.x()) * std::sin(rot.y()) * std::cos(rot.z()) - std::cos(rot.x()) * std::sin(rot.z()), std::cos(rot.x()) * std::sin(rot.y()) * std::cos(rot.z()) + std::sin(rot.x()) * std::sin(rot.z()), 0.0f,
+            std::cos(rot.y()) * std::sin(rot.z()), std::sin(rot.x()) * std::sin(rot.y()) * std::sin(rot.z()) + std::cos(rot.x()) * std::cos(rot.z()), std::cos(rot.x()) * std::sin(rot.y()) * std::sin(rot.z()) - std::sin(rot.x()) * std::cos(rot.z()), 0.0f,
+            -std::sin(rot.y()),                     std::sin(rot.x()) * std::cos(rot.y()),                                                                 std::cos(rot.x()) * std::cos(rot.y()),                                                                0.0f,
             0.0f, 0.0f, 0.0f, 1.0f
         });
